refactored user funcs
