+++ functions/package.json
@@ -14,9 +14,9 @@
     },
     "main": "index.js",
     "dependencies": {
-        "axios": "^1.6.0",
+        "axios": "^1.6.7",
         "firebase-admin": "^11.8.0",
-        "firebase-functions": "^4.6.0"
+        "firebase-functions": "^4.7.0"
     },
     "devDependencies": {
         "firebase-functions-test": "^3.1.0"
